fix：解决测试中的问题

+++ src/main/java/cn/jackbin/SimpleRecord/vo/RecordVO.java
@@ -19,7 +19,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class RecordVO {
     @Positive(message = "花费类别为整数")
-    private Long spendCategoryId;
+    private Integer spendCategoryId;
 
     @NotNull(message = "金额不能为空")
     private Double amount;
